* commit 'df1ec76d54db3a91759fc93f4d6b6828d4ec0f59':
  Core: Updated to 11.0.2

+++ src/server/game/Entities/Player/CollectionMgr.cpp
@@ -307,7 +307,7 @@ void CollectionMgr::CheckHeirloomUpgrades(Item* item)
         if (newItemId)
         {
             std::vector<uint32> const& fields = player->GetDynamicValues(PLAYER_DYNAMIC_FIELD_HEIRLOOMS);
-            uint16 offset = uint16(std::find(fields.begin(), fields.end(), int32(itr->first)) - fields.begin());
+            uint16 offset = uint16(std::find(fields.begin(), fields.end(), itr->first) - fields.begin());
 
             player->SetHeirloom(offset, newItemId);
             player->SetHeirloomFlags(offset, 0);
